fix up scripts, add readme, and install hint

+++ clenv.sh
@@ -1,10 +1,10 @@
 #!/usr/bin/env bash
 
-# this script creates a copy of the current system-level
-# StdEnv files of the Clean distribution and places them
-# in the users home directory (under .cache)
+# this script copies over the entire directory tree
+# of the clean project into your homedirectory.
 
 _install_dir="$HOME/.cache/clean-itasks-dev-bin"
+_source_dir=/opt/clean-itasks-dev-bin
 
 usage () {
     echo "USAGE: clenv [cmd]" >&2
@@ -19,32 +19,34 @@ fi
 
 case $1 in
     status)
-        if [ ! -d "$_install_dir/lib" ]; then
+        if [ ! -d "$_install_dir" ]; then
             echo "not initilised"
         else
             echo "initilised"
         fi
         ;;
     init)
-        if [ -d "$_install_dir/lib" ]; then
+        if [ -d "$_install_dir" ]; then
             echo -n "already initilised, overwrite? [Y/n]: "
             read -r yn
-            if [ "$yn" != "${yn#[Yy]}" ] ;then
-                rm -rf "$_install_dir"
-                cp -r /opt/clean-itasks-dev-bin/lib "$_install_dir/"
+            if [ -z "$yn" ] || [ "$yn" != "${yn#[Yy]}" ] ;then
+                rm -rf "${_install_dir:?}"/*
+                cp -r "${_source_dir:?}"/* "$_install_dir/"
                 echo "done"
+            else
+                echo "skipped"
             fi
         else
             [ -d "$_install_dir" ] || mkdir -p "$_install_dir"
-            cp -r /opt/clean-itasks-dev-bin/lib "$_install_dir/"
+            cp -r "${_source_dir:?}"/* "$_install_dir/"
             echo "done"
         fi
         ;;
     deinit)
-        if [ ! -d "$_install_dir/lib" ]; then
+        if [ ! -d "$_install_dir" ]; then
             echo "not initilised, nothing to do :("
         else
-            rm -rf "$_install_dir"
+            rm -rf "${_install_dir:?}"
             echo "done"
         fi
         ;;
